Simplify GCM interface state management.
Instead of allowing multile calls to |CRYPTO_gcm128_aad| and
multiple calls to the |_encrypt| or |_decrypt| functions, instead
require that the caller provide all the AAD at once and all the
data to encrypt/decrypt at once. In this way, we can avoid the need to
manage partial blocks of data within the |GCM128_CONTEXT| structure.